Initial Commit
Right now it only works 3 directories deep cuz that's all I needed. only argument is the path to the portable drive, i.e. S:\ and of course the icons have to be on the drive. might have to press ctrl+r to see changes, but this way you can plug it into another computer and your icons will still show up even if the drive letter is different